Update to concrete5lockpick.php from SH
Changed line 124 of concrete5lockpick.php to remove the / as this was causing issues with C5 installed in a sub folder (script not found or unable to stat). Keeping the / made the script look in the root for the file instead of in the folder. Please see http://squirrelhosting.co.uk/hosting-blog/hosting-blog-info.php?id=68 (bottom of guide) for more details.